Move useSteps hook to tools/hooks folder
