+++ pages/renew.tsx
@@ -21,7 +21,7 @@ import {
 import { Step, Steps } from 'chakra-ui-steps'; // Chakra UI Steps
 import Layout from '@components/applicant/Layout'; // Layout component
 import ReviewRequestField from '@components/applicant/renewals/ReviewRequestField'; // Field in Review Request section
-import useSteps from '@tools/components/Steps/useSteps'; // Custom hook for managing steps state
+import useSteps from '@tools/hooks/useSteps'; // Custom hook for managing steps state
 import IncompleteSectionAlert from '@components/applicant/renewals/IncompleteSectionAlert'; // Alert box for incomplete form section
 
 export default function Renew() {
